Module wayland_protocols_wlr::foreign_toplevel::v1::client::zwlr_foreign_toplevel_handle_v1

an opened toplevel

A zwlr_foreign_toplevel_handle_v1 object represents an opened toplevel window. Each app may have multiple opened toplevels.

Each toplevel has a list of outputs it is visible on, conveyed to the client with the output_enter and output_leave events.

Structs

ZwlrForeignToplevelHandleV1

an opened toplevel

Enums

Error

Event

Request

State

types of states on the toplevel

Constants

EVT_APP_ID_OPCODE

The wire opcode for this event

EVT_APP_ID_SINCE

The minimal object version supporting this event

EVT_CLOSED_OPCODE

The wire opcode for this event

EVT_CLOSED_SINCE

The minimal object version supporting this event

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_OUTPUT_ENTER_OPCODE

The wire opcode for this event

EVT_OUTPUT_ENTER_SINCE

The minimal object version supporting this event

EVT_OUTPUT_LEAVE_OPCODE

The wire opcode for this event

EVT_OUTPUT_LEAVE_SINCE

The minimal object version supporting this event

EVT_PARENT_OPCODE

The wire opcode for this event

EVT_PARENT_SINCE

The minimal object version supporting this event

EVT_STATE_OPCODE

The wire opcode for this event

EVT_STATE_SINCE

The minimal object version supporting this event

EVT_TITLE_OPCODE

The wire opcode for this event

EVT_TITLE_SINCE

The minimal object version supporting this event

REQ_ACTIVATE_OPCODE

The wire opcode for this request

REQ_ACTIVATE_SINCE

The minimal object version supporting this request

REQ_CLOSE_OPCODE

The wire opcode for this request

REQ_CLOSE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_SET_FULLSCREEN_OPCODE

The wire opcode for this request

REQ_SET_FULLSCREEN_SINCE

The minimal object version supporting this request

REQ_SET_MAXIMIZED_OPCODE

The wire opcode for this request

REQ_SET_MAXIMIZED_SINCE

The minimal object version supporting this request

REQ_SET_MINIMIZED_OPCODE

The wire opcode for this request

REQ_SET_MINIMIZED_SINCE

The minimal object version supporting this request

REQ_SET_RECTANGLE_OPCODE

The wire opcode for this request

REQ_SET_RECTANGLE_SINCE

The minimal object version supporting this request

REQ_UNSET_FULLSCREEN_OPCODE

The wire opcode for this request

REQ_UNSET_FULLSCREEN_SINCE

The minimal object version supporting this request

REQ_UNSET_MAXIMIZED_OPCODE

The wire opcode for this request

REQ_UNSET_MAXIMIZED_SINCE

The minimal object version supporting this request

REQ_UNSET_MINIMIZED_OPCODE

The wire opcode for this request

REQ_UNSET_MINIMIZED_SINCE

The minimal object version supporting this request