Module wayland_protocols_wlr::layer_shell::v1::server::zwlr_layer_surface_v1

layer metadata interface

An interface that may be implemented by a wl_surface, for surfaces that are designed to be rendered as a layer of a stacked desktop-like environment.

Layer surface state (layer, size, anchor, exclusive zone, margin, interactivity) is double-buffered, and will be applied at the time wl_surface.commit of the corresponding wl_surface is called.

Attaching a null buffer to a layer surface unmaps it.

Unmapping a layer_surface means that the surface cannot be shown by the compositor until it is explicitly mapped again. The layer_surface returns to the state it had right after layer_shell.get_layer_surface. The client can re-map the surface by performing a commit without any buffer attached, waiting for a configure event and handling it as usual.

Structs

Anchor

ZwlrLayerSurfaceV1

layer metadata interface

Enums

Error

Event

KeyboardInteractivity

types of keyboard interaction possible for a layer shell surface

Request

Constants

EVT_CLOSED_OPCODE

The wire opcode for this event

EVT_CLOSED_SINCE

The minimal object version supporting this event

EVT_CONFIGURE_OPCODE

The wire opcode for this event

EVT_CONFIGURE_SINCE

The minimal object version supporting this event

REQ_ACK_CONFIGURE_OPCODE

The wire opcode for this request

REQ_ACK_CONFIGURE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_GET_POPUP_OPCODE

The wire opcode for this request

REQ_GET_POPUP_SINCE

The minimal object version supporting this request

REQ_SET_ANCHOR_OPCODE

The wire opcode for this request

REQ_SET_ANCHOR_SINCE

The minimal object version supporting this request

REQ_SET_EXCLUSIVE_ZONE_OPCODE

The wire opcode for this request

REQ_SET_EXCLUSIVE_ZONE_SINCE

The minimal object version supporting this request

REQ_SET_KEYBOARD_INTERACTIVITY_OPCODE

The wire opcode for this request

REQ_SET_KEYBOARD_INTERACTIVITY_SINCE

The minimal object version supporting this request

REQ_SET_LAYER_OPCODE

The wire opcode for this request

REQ_SET_LAYER_SINCE

The minimal object version supporting this request

REQ_SET_MARGIN_OPCODE

The wire opcode for this request

REQ_SET_MARGIN_SINCE

The minimal object version supporting this request

REQ_SET_SIZE_OPCODE

The wire opcode for this request

REQ_SET_SIZE_SINCE

The minimal object version supporting this request