Module wayland_protocols_wlr::output_management::v1::server::zwlr_output_manager_v1

output device configuration manager

This interface is a manager that allows reading and writing the current output device configuration.

Output devices that display pixels (e.g. a physical monitor or a virtual output in a window) are represented as heads. Heads cannot be created nor destroyed by the client, but they can be enabled or disabled and their properties can be changed. Each head may have one or more available modes.

Whenever a head appears (e.g. a monitor is plugged in), it will be advertised via the head event. Immediately after the output manager is bound, all current heads are advertised.

Whenever a head’s properties change, the relevant wlr_output_head events will be sent. Not all head properties will be sent: only properties that have changed need to.

Whenever a head disappears (e.g. a monitor is unplugged), a wlr_output_head.finished event will be sent.

After one or more heads appear, change or disappear, the done event will be sent. It carries a serial which can be used in a create_configuration request to update heads properties.

The information obtained from this protocol should only be used for output configuration purposes. This protocol is not designed to be a generic output property advertisement protocol for regular clients. Instead, protocols such as xdg-output should be used.

Structs

ZwlrOutputManagerV1

output device configuration manager

Enums

Event

Request

Constants

EVT_DONE_OPCODE

The wire opcode for this event

EVT_DONE_SINCE

The minimal object version supporting this event

EVT_FINISHED_OPCODE

The wire opcode for this event

EVT_FINISHED_SINCE

The minimal object version supporting this event

EVT_HEAD_OPCODE

The wire opcode for this event

EVT_HEAD_SINCE

The minimal object version supporting this event

REQ_CREATE_CONFIGURATION_OPCODE

The wire opcode for this request

REQ_CREATE_CONFIGURATION_SINCE

The minimal object version supporting this request

REQ_STOP_OPCODE

The wire opcode for this request

REQ_STOP_SINCE

The minimal object version supporting this request