wayland_protocols_wlr::screencopy::v1::client::zwlr_screencopy_frame_v1
Enum Event

#[non_exhaustive]
pub enum Event {
    Buffer {
        format: WEnum<Format>,
        width: u32,
        height: u32,
        stride: u32,
    },
    Flags {
        flags: WEnum<Flags>,
    },
    Ready {
        tv_sec_hi: u32,
        tv_sec_lo: u32,
        tv_nsec: u32,
    },
    Failed,
    Damage {
        x: u32,
        y: u32,
        width: u32,
        height: u32,
    },
    LinuxDmabuf {
        format: u32,
        width: u32,
        height: u32,
    },
    BufferDone,
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Buffer

wl_shm buffer information

Provides information about wl_shm buffer parameters that need to be used for this frame. This event is sent once after the frame is created if wl_shm buffers are supported.

Fields

format: WEnum<Format>

buffer format

width: u32

buffer width

height: u32

buffer height

stride: u32

buffer stride

Flags

frame flags

Provides flags about the frame. This event is sent once before the “ready” event.

Fields

flags: WEnum<Flags>

frame flags

Ready

indicates frame is available for reading

Called as soon as the frame is copied, indicating it is available for reading. This event includes the time at which presentation happened at.

The timestamp is expressed as tv_sec_hi, tv_sec_lo, tv_nsec triples, each component being an unsigned 32-bit value. Whole seconds are in tv_sec which is a 64-bit value combined from tv_sec_hi and tv_sec_lo, and the additional fractional part in tv_nsec as nanoseconds. Hence, for valid timestamps tv_nsec must be in [0, 999999999]. The seconds part may have an arbitrary offset at start.

After receiving this event, the client should destroy the object.

Fields

tv_sec_hi: u32

high 32 bits of the seconds part of the timestamp

tv_sec_lo: u32

low 32 bits of the seconds part of the timestamp

tv_nsec: u32

nanoseconds part of the timestamp

Failed

frame copy failed

This event indicates that the attempted frame copy has failed.

After receiving this event, the client should destroy the object.

Damage

carries the coordinates of the damaged region

This event is sent right before the ready event when copy_with_damage is requested. It may be generated multiple times for each copy_with_damage request.

The arguments describe a box around an area that has changed since the last copy request that was derived from the current screencopy manager instance.

The union of all regions received between the call to copy_with_damage and a ready event is the total damage since the prior ready event.

Only available since version 2 of the interface

Fields

x: u32

damaged x coordinates

y: u32

damaged y coordinates

width: u32

current width

height: u32

current height

LinuxDmabuf

linux-dmabuf buffer information

Provides information about linux-dmabuf buffer parameters that need to be used for this frame. This event is sent once after the frame is created if linux-dmabuf buffers are supported.

Only available since version 3 of the interface

Fields

format: u32

fourcc pixel format

width: u32

buffer width

height: u32

buffer height

BufferDone

all buffer types reported

This event is sent once after all buffer events have been sent.

The client should proceed to create a buffer of one of the supported types, and send a “copy” request.

Only available since version 3 of the interface

Implementations

impl Event

pub fn opcode(&self) -> u16

Get the opcode number of this message