wayland_protocols_wlr::virtual_pointer::v1::client

Module zwlr_virtual_pointer_v1

virtual pointer

This protocol allows clients to emulate a physical pointer device. The requests are mostly mirror opposites of those specified in wl_pointer.

Structs

ZwlrVirtualPointerV1

virtual pointer

Enums

Error

Event

Request

Constants

REQ_AXIS_DISCRETE_OPCODE

The wire opcode for this request

REQ_AXIS_DISCRETE_SINCE

The minimal object version supporting this request

REQ_AXIS_OPCODE

The wire opcode for this request

REQ_AXIS_SINCE

The minimal object version supporting this request

REQ_AXIS_SOURCE_OPCODE

The wire opcode for this request

REQ_AXIS_SOURCE_SINCE

The minimal object version supporting this request

REQ_AXIS_STOP_OPCODE

The wire opcode for this request

REQ_AXIS_STOP_SINCE

The minimal object version supporting this request

REQ_BUTTON_OPCODE

The wire opcode for this request

REQ_BUTTON_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request

REQ_FRAME_OPCODE

The wire opcode for this request

REQ_FRAME_SINCE

The minimal object version supporting this request

REQ_MOTION_ABSOLUTE_OPCODE

The wire opcode for this request

REQ_MOTION_ABSOLUTE_SINCE

The minimal object version supporting this request

REQ_MOTION_OPCODE

The wire opcode for this request

REQ_MOTION_SINCE

The minimal object version supporting this request