Crate wayland_protocols_wlr

This crate provides bindings to the wlroots wayland protocol extensions provided in https://gitlab.freedesktop.org/wlroots/wlr-protocols

These bindings are built on top of the crates wayland-client and wayland-server.

Each protocol module contains a client and a server submodules, for each side of the protocol. The creation of these modules (and the dependency on the associated crate) is controlled by the two cargo features client and server.

Modules

data_control

Control data devices, particularly the clipboard.

export_dmabuf

A protocol for low overhead screen content capturing

foreign_toplevel

List and control opened apps

gamma_control

Manage gamma tables of outputs.

input_inhibitor

Inhibits input events to other clients

layer_shell

Layered shell protocol

output_management

Output management protocol

output_power_management

Output power management protocol

screencopy

Screen content capturing on client buffers

virtual_pointer

Virtual pointer protocol