Module wayland_protocols_wlr::export_dmabuf::v1::client::zwlr_export_dmabuf_frame_v1

a DMA-BUF frame

This object represents a single DMA-BUF frame.

If the capture is successful, the compositor will first send a “frame” event, followed by one or several “object”. When the frame is available for readout, the “ready” event is sent.

If the capture failed, the “cancel” event is sent. This can happen anytime before the “ready” event.

Once either a “ready” or a “cancel” event is received, the client should destroy the frame. Once an “object” event is received, the client is responsible for closing the associated file descriptor.

All frames are read-only and may not be written into or altered.

Structs

ZwlrExportDmabufFrameV1

a DMA-BUF frame

Enums

CancelReason

cancel reason

Event

Flags

frame flags

Request

Constants

EVT_CANCEL_OPCODE

The wire opcode for this event

EVT_CANCEL_SINCE

The minimal object version supporting this event

EVT_FRAME_OPCODE

The wire opcode for this event

EVT_FRAME_SINCE

The minimal object version supporting this event

EVT_OBJECT_OPCODE

The wire opcode for this event

EVT_OBJECT_SINCE

The minimal object version supporting this event

EVT_READY_OPCODE

The wire opcode for this event

EVT_READY_SINCE

The minimal object version supporting this event

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request