Module wayland_protocols_wlr::output_management::v1::client::zwlr_output_head_v1

output device

A head is an output device. The difference between a wl_output object and a head is that heads are advertised even if they are turned off. A head object only advertises properties and cannot be used directly to change them.

A head has some read-only properties: modes, name, description and physical_size. These cannot be changed by clients.

Other properties can be updated via a wlr_output_configuration object.

Properties sent via this interface are applied atomically via the wlr_output_manager.done event. No guarantees are made regarding the order in which properties are sent.

Structs

ZwlrOutputHeadV1

output device

Enums

AdaptiveSyncState

Event

Request

Constants

EVT_ADAPTIVE_SYNC_OPCODE

The wire opcode for this event

EVT_ADAPTIVE_SYNC_SINCE

The minimal object version supporting this event

EVT_CURRENT_MODE_OPCODE

The wire opcode for this event

EVT_CURRENT_MODE_SINCE

The minimal object version supporting this event

EVT_DESCRIPTION_OPCODE

The wire opcode for this event

EVT_DESCRIPTION_SINCE

The minimal object version supporting this event

EVT_ENABLED_OPCODE

The wire opcode for this event

EVT_ENABLED_SINCE

The minimal object version supporting this event

EVT_FINISHED_OPCODE

The wire opcode for this event

EVT_FINISHED_SINCE

The minimal object version supporting this event

EVT_MAKE_OPCODE

The wire opcode for this event

EVT_MAKE_SINCE

The minimal object version supporting this event

EVT_MODEL_OPCODE

The wire opcode for this event

EVT_MODEL_SINCE

The minimal object version supporting this event

EVT_MODE_OPCODE

The wire opcode for this event

EVT_MODE_SINCE

The minimal object version supporting this event

EVT_NAME_OPCODE

The wire opcode for this event

EVT_NAME_SINCE

The minimal object version supporting this event

EVT_PHYSICAL_SIZE_OPCODE

The wire opcode for this event

EVT_PHYSICAL_SIZE_SINCE

The minimal object version supporting this event

EVT_POSITION_OPCODE

The wire opcode for this event

EVT_POSITION_SINCE

The minimal object version supporting this event

EVT_SCALE_OPCODE

The wire opcode for this event

EVT_SCALE_SINCE

The minimal object version supporting this event

EVT_SERIAL_NUMBER_OPCODE

The wire opcode for this event

EVT_SERIAL_NUMBER_SINCE

The minimal object version supporting this event

EVT_TRANSFORM_OPCODE

The wire opcode for this event

EVT_TRANSFORM_SINCE

The minimal object version supporting this event

REQ_RELEASE_OPCODE

The wire opcode for this request

REQ_RELEASE_SINCE

The minimal object version supporting this request