Module wayland_protocols_wlr::screencopy::v1::client::zwlr_screencopy_frame_v1

a frame ready for copy

This object represents a single frame.

When created, a series of buffer events will be sent, each representing a supported buffer type. The “buffer_done” event is sent afterwards to indicate that all supported buffer types have been enumerated. The client will then be able to send a “copy” request. If the capture is successful, the compositor will send a “flags” followed by a “ready” event.

For objects version 2 or lower, wl_shm buffers are always supported, ie. the “buffer” event is guaranteed to be sent.

If the capture failed, the “failed” event is sent. This can happen anytime before the “ready” event.

Once either a “ready” or a “failed” event is received, the client should destroy the frame.

Structs

Flags

ZwlrScreencopyFrameV1

a frame ready for copy

Enums

Error

Event

Request

Constants

EVT_BUFFER_DONE_OPCODE

The wire opcode for this event

EVT_BUFFER_DONE_SINCE

The minimal object version supporting this event

EVT_BUFFER_OPCODE

The wire opcode for this event

EVT_BUFFER_SINCE

The minimal object version supporting this event

EVT_DAMAGE_OPCODE

The wire opcode for this event

EVT_DAMAGE_SINCE

The minimal object version supporting this event

EVT_FAILED_OPCODE

The wire opcode for this event

EVT_FAILED_SINCE

The minimal object version supporting this event

EVT_FLAGS_OPCODE

The wire opcode for this event

EVT_FLAGS_SINCE

The minimal object version supporting this event

EVT_LINUX_DMABUF_OPCODE

The wire opcode for this event

EVT_LINUX_DMABUF_SINCE

The minimal object version supporting this event

EVT_READY_OPCODE

The wire opcode for this event

EVT_READY_SINCE

The minimal object version supporting this event

REQ_COPY_OPCODE

The wire opcode for this request

REQ_COPY_SINCE

The minimal object version supporting this request

REQ_COPY_WITH_DAMAGE_OPCODE

The wire opcode for this request

REQ_COPY_WITH_DAMAGE_SINCE

The minimal object version supporting this request

REQ_DESTROY_OPCODE

The wire opcode for this request

REQ_DESTROY_SINCE

The minimal object version supporting this request